views- aplicação

+++ pokepoint/app/app_views/checkin.py
@@ -11,7 +11,11 @@ from django.contrib import messages
 
 @login_required(login_url='login')
 def listCheckin(request):
-    check_list = CheckIn.objects.all()
+    is_manager = request.user.groups.filter(name='manager').exists()
+    if is_manager:
+        check_list = CheckIn.objects.all()
+    else:
+        check_list = CheckIn.objects.filter(timeCard__employee=request.user)
     template_name = 'app/checkin.html'
     return render(request, template_name, {'checkList': check_list})
 
@@ -31,21 +35,29 @@ def addCheckin(request):
 
 @login_required(login_url='login')
 def editCheckin(request, pk):
-    checkin = CheckIn.objects.get(id=pk)
-    form = CheckinForm(instance=checkin)
-    template_name = 'app/form.html'
-    page_name = 'Add Checkin'
-    if request.method == 'POST':
-        form = CheckinForm(request.POST, instance=checkin)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Check-in Edited')
-            return redirect('list_checkin')
-    return render(request, template_name, {'form': form, 'pageName': page_name})
+    checkin = CheckIn.objects.filter(id=pk, timeCard__employee=request.user)
+    is_manager = request.user.groups.filter(name='manager').exists()
+    if is_manager or checkin:
+        form = CheckinForm(instance=checkin)
+        template_name = 'app/form.html'
+        page_name = 'Add Checkin'
+        if request.method == 'POST':
+            form = CheckinForm(request.POST, instance=checkin)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Check-in Edited')
+                return redirect('list_checkin')
+        return render(request, template_name, {'form': form, 'pageName': page_name})
+    messages.error(request, 'Não tem permissão para aceder a pagina')
+    return redirect('list_checkin')
 
 
 @login_required(login_url='login')
 def deleteCheckin(request, pk):
-    checkin = get_object_or_404(CheckIn, pk=pk)
-    checkin.delete()
+    is_manager = request.user.groups.filter(name='manager').exists()
+    if is_manager:
+        checkin = get_object_or_404(CheckIn, pk=pk)
+        checkin.delete()
+        return redirect('list_checkin')
+    messages.error(request, 'Não tem permissão para aceder a pagina')
     return redirect('list_checkin')
